:package: Move redux/store related code into own folder.
We previously adhered to grouping stuff by domain. This meant that
reducer/actions/selectors for the store would be mixed inside the
`containers` folder with JSX components.

This "coupled" the store to a feature/domain ...kinda. We're trying
to decouple the store completely from the view layer of our app and
moved everything to a separate `store` folder, that resembles the
internal app state structure.